style: update base 64 encoder page buttons to be more readable (#53)
Updated the format of the buttons to be more similar to the rest of the
site.
Used the endIcon property instead of putting the icon inside. Also gave
the encode and decode buttons text to be more readable.
Fix for #50.
Update copy button to use ContentCopy rather than AssignmentTurnedIn to
be more understandable and intuitive.